refactor: add ShareComponent
