+++ app/components/LikesCounter.jsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { useCallback, useEffect, useRef, useState } from 'react'
+import { useCallback, useEffect, useState } from 'react'
 
 import { usePathname, useRouter } from 'next/navigation'
 
@@ -8,24 +8,13 @@ import { Image } from '@nextui-org/image'
 
 import { toggleLike as toggleLikeAction } from '../(web)/[@username]/actions/imagesActions'
 import { BVButton } from './BVButton'
-import ShareDropdown from './ShareDropdown'
 
 function LikesCounter({ imageInfo, isLike, isAuthenticated, totalLikes }) {
   const [isLiked, setIsLiked] = useState(!!isLike)
   const [count, setCount] = useState(totalLikes)
   const [isLoading, setIsLoading] = useState(false)
-  const [isDropdownOpen, setIsDropdownOpen] = useState(false)
-  const [currentUrl, setCurrentUrl] = useState('')
-  const dropdownRef = useRef(null)
-
   const pathname = usePathname()
   const router = useRouter()
-
-  useEffect(() => {
-    if (typeof window !== 'undefined') {
-      setCurrentUrl(window.location.href)
-    }
-  }, [])
 
   const handleToggleLike = useCallback(() => {
     setIsLiked((prevIsLiked) => !prevIsLiked)
@@ -62,26 +51,10 @@ function LikesCounter({ imageInfo, isLike, isAuthenticated, totalLikes }) {
     }
   }
 
-  const handleShare = useCallback(() => {
-    setIsDropdownOpen((prev) => !prev)
-  }, [])
-
-  useEffect(() => {
-    const handleClickOutside = (event) => {
-      if (dropdownRef.current && !dropdownRef.current.contains(event.target)) {
-        setIsDropdownOpen(false)
-      }
-    }
-    document.addEventListener('mousedown', handleClickOutside)
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside)
-    }
-  }, [])
-
   const commonButtonClasses = 'w-1/2 leading-4'
 
   return (
-    <div className="relative flex items-center gap-2.5">
+    <>
       <BVButton
         size="xl"
         variant="bordered"
@@ -101,25 +74,7 @@ function LikesCounter({ imageInfo, isLike, isAuthenticated, totalLikes }) {
       >
         {count}
       </BVButton>
-
-      <BVButton
-        size="xl"
-        variant="bordered"
-        color="background"
-        aria-label="Share"
-        className={commonButtonClasses}
-        onClick={handleShare}
-        isIconOnly
-      >
-        <Image src="/share.svg" alt="share" radius="none" width={20} height={20} />
-      </BVButton>
-
-      <ShareDropdown
-        isDropdownOpen={isDropdownOpen}
-        setIsDropdownOpen={setIsDropdownOpen}
-        currentUrl={currentUrl}
-      />
-    </div>
+    </>
   )
 }
